html/semantics/forms/form-submission-0/form-data-set-usv.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Submitting a form data set that contains unpaired surrogates must convert to Unicode scalar values</title>
<link rel="author" title="Domenic Denicola" href="mailto:d@domenic.me">
<link rel="help" href="https://html.spec.whatwg.org/multipage/forms.html#constructing-form-data-set">
<link rel="help" href="https://github.com/whatwg/html/issues/1490">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<iframe id="testframe" src="form-data-set-usv-form.html"></iframe>

<script>
"use strict";

async_test(t => {
  window.onload = t.step_func(() => {
    const iframe = document.querySelector("#testframe");
    const form = iframe.contentWindow.document.querySelector("form");

    iframe.onload = t.step_func_done(() => {
      const result = iframe.contentWindow.document.body.textContent;

      // trim() since there are unrelated bugs in Chrome (which does not append \r\n when doing
      // text/plain serialization)
      assert_equals(result.trim(), `input1\uFFFD=input1\uFFFD
input2\uFFFD=input2\uFFFD
input3\uFFFD=input3\uFFFD
input4\uFFFD=input4\uFFFD`);
    });

    form.submit();
  });
});
</script>
